Modify fknum type to smalling for phonefunckey [bugfix] Remove IF EXISTS on SQLite schema
git-svn-id: https://rcs.lan.proformatique.com/svn/xivo/trunk@2260 dc85baef-7018-0410-bfbd-8a639da648bb

+++ base-config/asterisk/astmysql.db.sql
@@ -768,7 +768,7 @@ CREATE UNIQUE INDEX `phonebooknumber__uidx__phonebookid_type` ON `phonebooknumbe
 DROP TABLE IF EXISTS `phonefunckey`;
 CREATE TABLE `phonefunckey` (
  `iduserfeatures` int(10) unsigned NOT NULL,
- `fknum` int(10) unsigned NOT NULL,
+ `fknum` smallint unsigned NOT NULL,
  `exten` varchar(40),
  `typeextenumbers` enum('extenfeatures','featuremap','generalfeatures'),
  `typevalextenumbers` varchar(255),
